crypto-trade-simulator/requirements.txt
websockets
pandas
numpy
scikit-learn
statsmodels
PyQt5
matplotlib
